+++ examples/main/modules.tf
@@ -91,7 +91,7 @@ module "linux_scaleset" {
   admin_username = "myusername"
   ssh_public_key = var.ssh_public_key
 
-  vms_sku = "Standard_B2s"
+  vms_size = "Standard_B2s"
 
   subnet_id = module.subnet.subnet_id