+++ variables.tf
@@ -1,153 +1,153 @@
 variable "resource_group_name" {
-  description = "Name of the resource group"
+  description = "Name of the resource group."
   type        = string
 }
 
 variable "location" {
-  description = "Azure region to use"
+  description = "Azure region to use."
   type        = string
 }
 
 variable "location_short" {
-  description = "Short string for Azure location"
+  description = "Short string for Azure location."
   type        = string
 }
 
 variable "environment" {
-  description = "Project environment"
+  description = "Project environment."
   type        = string
 }
 
 variable "stack" {
-  description = "Project stack name"
+  description = "Project stack name."
   type        = string
 }
 
 variable "client_name" {
-  description = "Client name/account used in naming"
+  description = "Client name/account used in naming."
   type        = string
 }
 
 variable "ssh_private_key" {
-  description = "Private SSH key deployed on Scale set"
+  description = "Private SSH key deployed on Scale set."
   type        = string
   default     = null
 }
 
 variable "ssh_public_key" {
-  description = "Public SSH key deployed on Scale set"
+  description = "Public SSH key deployed on Scale set."
   type        = string
   default     = null
 }
 
 variable "accelerated_networking" {
-  description = "Specifies whether to enable accelerated networking or not"
+  description = "Whether to enable accelerated networking or not."
   type        = bool
-  default     = false
+  default     = true
 }
 
 variable "dns_servers" {
-  description = "Specifies an array of DNS servers"
+  description = "List of DNS servers."
   type        = list(string)
   default     = []
 }
 
-variable "ip_forwarding" {
-  description = "Whether IP forwarding is enabled on this NIC"
+variable "ip_forwarding_enabled" {
+  description = "Whether IP forwarding is enabled on this NIC."
   type        = bool
   default     = false
 }
 
 variable "network_security_group_id" {
-  description = "Specifies the id for the network security group"
+  description = "ID of the Network Security Group."
   type        = string
-  default     = ""
+  default     = null
 }
 
 variable "subnet_id" {
-  description = "Specifies the identifier of the subnet"
+  description = "ID of the subnet"
   type        = string
 }
 
 variable "application_gateway_backend_address_pool_ids" {
-  description = "Specifies an array of references to backend address pools of application gateways. A scale set can reference backend address pools of one application gateway"
+  description = "List of references to backend address pools of Application Gateways. A Scale Set can reference backend address pools of one Application Gateway."
   type        = list(string)
   default     = []
 }
 
 variable "load_balancer_backend_address_pool_ids" {
-  description = "Specifies an array of references to backend address pools of load balancers. A scale set can reference backend address pools of one public and one internal load balancer"
+  description = "List of references to backend address pools of Load Balancers. A Scale Set can reference backend address pools of one public and one internal Load Balancer."
   type        = list(string)
   default     = []
 }
 
 variable "load_balancer_inbound_nat_rules_ids" {
-  description = "Specifies an array of references to inbound NAT rules for load balancers"
+  description = "List of references to inbound NAT rules for Load Balancers."
   type        = list(string)
   default     = []
 }
 
 variable "application_security_group_ids" {
-  description = "Specifies up to 20 application security group IDs"
+  description = "IDs of Application Security Group IDs (up to 20)."
   type        = list(string)
   default     = []
 }
 
-variable "vms_sku" {
-  description = "Specifies the size of virtual machines in a scale set"
+variable "vms_size" {
+  description = "Size (SKU) of Virtual Machines in a Scale Set."
   type        = string
 }
 
 variable "os_disk_managed_disk_type" {
-  description = "Specifies the type of managed disk to create [Possible values : Standard_LRS, StandardSSD_LRS or Premium_LRS]"
+  description = "Type of managed disk to create [Possible values : Standard_LRS, StandardSSD_LRS or Premium_LRS]."
   type        = string
-  default     = "Standard_LRS"
+  default     = "StandardSSD_LRS"
 }
 
 variable "os_disk_caching" {
-  description = "Specifies the caching requirements [Possible values : None, ReadOnly, ReadWrite]"
+  description = "OS disk caching requirements [Possible values : None, ReadOnly, ReadWrite]."
   type        = string
   default     = "None"
 }
 
 variable "os_disk_size_gb" {
-  description = "Size of the OS disk in GB"
+  description = "Size of the OS disk in GB."
   type        = number
   default     = 32
 }
 
-variable "os_disk_is_local" {
-  description = "Specifies the Ephemeral Disk Settings for the OS Disk to Local"
+variable "os_ephemeral_disk_enabled" {
+  description = "Whether OS disk is local ephemeral disk. See https://learn.microsoft.com/en-us/azure/virtual-machines/ephemeral-os-disks."
   type        = bool
-  default     = false
+  default     = true
 }
 
 variable "os_disk_encryption_set_id" {
-  description = "The ID of the Disk Encryption Set which should be used to encrypt this Data Disk"
+  description = "ID of the Disk Encryption Set which should be used to encrypt the OS disk."
   type        = string
   default     = null
 }
 
 variable "os_disk_write_accelerator_enabled" {
-  description = "True to enable Write Accelerator for this Data Disk"
+  description = "Whether to enable write accelerator for the OS disk."
   type        = bool
   default     = false
 }
 
 variable "automatic_os_upgrade" {
-  description = "Automatic OS patches can be applied by Azure to your scaleset. This is particularly useful when upgrade_policy_mode is set to Rolling."
+  description = "Whether if automatic OS patches can be applied by Azure to your Scale Set. This is particularly useful when upgrade_policy_mode is set to Rolling."
   type        = bool
   default     = false
 }
 
 variable "disable_automatic_rollback" {
-  description = "Disable automatic rollback in case of failured"
+  description = "Disable automatic rollback in case of failures."
   type        = bool
   default     = false
 }
 
 variable "rolling_upgrade_policy" {
-  description = "This is only applicable when the upgrade_policy_mode is Rolling."
+  description = "Rolling upgrade policy, only applicable when the upgrade_policy_mode is Rolling."
   type = object({
     max_batch_instance_percent              = number
     max_unhealthy_instance_percent          = number
@@ -163,60 +163,81 @@ variable "rolling_upgrade_policy" {
 }
 
 variable "automatic_instance_repair" {
-  description = "Enable automatic instance repair. Must have health_probe_id or an Application Health Extension"
+  description = "Whether to enable automatic instance repair. Must have health_probe_id or an Application Health Extension"
   type        = bool
   default     = false
 }
 
 variable "health_probe_id" {
-  description = "Specifies the identifier for the load balancer health probe. Required when using Rolling as your upgrade_policy_mode."
+  description = "Specifies the identifier for the Load Balancer health probe. Required when using Rolling as your upgrade_policy_mode."
   type        = string
   default     = null
 }
 
 variable "boot_diagnostics_storage_uri" {
-  description = "Blob endpoint for the storage account to hold the virtual machine's diagnostic files"
+  description = "Blob endpoint for the Storage Account to hold the Virtual Machines diagnostic files."
   type        = string
   default     = ""
 }
 
 variable "extensions" {
-  description = "Can be specified to add extension profiles to the scale set"
-  type        = map(any)
-  default     = {}
+  description = "Extension profiles to add to the Scale Set."
+  type = list(object({
+    name                       = string
+    publisher                  = string
+    type                       = string
+    type_handler_version       = string
+    auto_upgrade_minor_version = optional(bool, true)
+    force_update_tag           = optional(string, null)
+    protected_settings         = optional(string, null)
+    provision_after_extensions = optional(list(string), [])
+    settings                   = optional(string, null)
+  }))
+  default = []
 }
 
 variable "data_disks" {
-  description = "A storage profile data disk"
-  type        = list(any)
-  default     = []
+  description = "Data disks profiles to attach."
+  type = list(object({
+    name                      = string
+    lun                       = number
+    disk_size_gb              = optional(number, null)
+    create_option             = optional(string, "Empty")
+    caching                   = optional(string, "None")
+    storage_account_type      = optional(string, "StandardSSD_LRS")
+    disk_encryption_set_id    = optional(string, null)
+    disk_iops_read_write      = optional(string, null)
+    disk_mbps_read_write      = optional(string, null)
+    write_accelerator_enabled = optional(string, null)
+  }))
+  default = []
 }
 
 variable "ultra_ssd_enabled" {
-  description = "Should the capacity to enable Data Disks of the UltraSSD_LRS storage account type be supported on this Virtual Machine Scale Set?"
+  description = "Whether UltraSSD_LRS storage account type can be enabled."
   type        = bool
   default     = false
 }
 
 variable "admin_username" {
-  description = "Username to use as admin user"
+  description = "Username of the administrator account of the Virtual Machines."
   type        = string
 }
 
 variable "admin_password" {
-  description = "Password for the administrator account of the virtual machine."
+  description = "Password for the administrator account of the Virtual Machines."
   type        = string
   default     = null
 }
 
 variable "instances_count" {
-  description = "Specify the number of instances to run"
+  description = "Number of instances in the Scale Set."
   type        = number
-  default     = 1
+  default     = 2
 }
 
 variable "source_image_reference" {
-  description = "Source Image references"
+  description = "Virtual Machines source image reference."
   type = object({
     publisher = string
     offer     = string
@@ -227,7 +248,7 @@ variable "source_image_reference" {
 }
 
 variable "source_image_id" {
-  description = "Id of the image to use."
+  description = "ID of the Virtual Machines image to use."
   type        = string
   default     = null
 }
@@ -245,7 +266,7 @@ variable "user_data" {
 }
 
 variable "scale_in_policy" {
-  description = "The scale-in policy rule that decides which virtual machines are chosen for removal when a Virtual Machine Scale Set is scaled in. Possible values for the scale-in policy rules are Default, NewestVM and OldestVM, defaults to Default"
+  description = "The scale-in policy rule that decides which Virtual Machines are chosen for removal when a Virtual Machine Scale Set is scaled in. Possible values for the scale-in policy rules are Default, NewestVM and OldestVM, defaults to Default."
   type        = string
   default     = "Default"
 }
@@ -262,8 +283,8 @@ variable "upgrade_mode" {
   default     = "Manual"
 }
 
-variable "zone_balance" {
-  description = "Should the Virtual Machines in this Scale Set be strictly evenly distributed across Availability Zones? Changing this forces a new resource to be created."
+variable "zone_balancing_enabled" {
+  description = "Whether the Virtual Machines in this Scale Set should be strictly evenly distributed across Availability Zones? Changing this forces a new resource to be created."
   type        = bool
   default     = true
 }
@@ -275,7 +296,7 @@ variable "zones_list" {
 }
 
 variable "identity" {
-  description = "Map with identity block informations as described here https://www.terraform.io/docs/providers/azurerm/r/linux_virtual_machine_scale_set.html#identity"
+  description = "Identity block information as described here https://www.terraform.io/docs/providers/azurerm/r/linux_virtual_machine_scale_set.html#identity."
   type = object({
     type         = string
     identity_ids = list(string)